Abstract pixels searching into component
